Remove paasta_execute_docker_command script, but copy execute_in_container function to local_run which still uses it.
