+++ setup.py
@@ -62,7 +62,6 @@ setup(
         "paasta_tools/kubernetes/bin/paasta_secrets_sync.py",
         "paasta_tools/log_task_lifecycle_events.py",
         "paasta_tools/paasta_deploy_tron_jobs",
-        "paasta_tools/paasta_execute_docker_command.py",
         "paasta_tools/paasta_remote_run.py",
         "paasta_tools/setup_istio_mesh.py",
         "paasta_tools/setup_kubernetes_cr.py",
